Refactored console to make it easier to ignore the demo stuff
